move ArrayToYieldDataProvider to PHPUnit group
