--- config/level/architecture/phpunit.yml
@@ -1,2 +1,0 @@
-services:
-    Rector\Rector\Architecture\PHPUnit\ArrayToYieldDataProviderRector: ~
