+++ config/level/phpunit/phpunit-yield-data-provider.yml
@@ -1,0 +1,2 @@
+services:
+    Rector\PHPUnit\Rector\ArrayToYieldDataProviderRector: ~
