+++ packages/PHPUnit/tests/Rector/ArrayToYieldDataProviderRector/Correct/correct3.php.inc
