+++ packages/PHPUnit/tests/Rector/ArrayToYieldDataProviderRector/config.yml
@@ -1,0 +1,2 @@
+services:
+    Rector\PHPUnit\Rector\ArrayToYieldDataProviderRector: ~
