--- tests/Rector/Architecture/PHPUnit/config.yml
@@ -1,2 +1,0 @@
-services:
-    Rector\Rector\Architecture\PHPUnit\ArrayToYieldDataProviderRector: ~
